+++ .swift-format
@@ -1,58 +1,64 @@
 {
-  "fileScopedDeclarationPrivacy" : {
-    "accessLevel" : "private"
+  "version": 1,
+  "lineLength": 80,
+  "maximumBlankLines": 1,
+  "fileScopedDeclarationPrivacy": {
+    "accessLevel": "private"
   },
-  "indentation" : {
-    "spaces" : 4
+  "tabWidth": 4,
+  "indentation": {
+    "spaces": 4
   },
-  "indentConditionalCompilationBlocks" : false,
-  "indentSwitchCaseLabels" : false,
-  "lineBreakAroundMultilineExpressionChainComponents" : true,
-  "lineBreakBeforeControlFlowKeywords" : true,
-  "lineBreakBeforeEachArgument" : true,
-  "lineBreakBeforeEachGenericRequirement" : true,
-  "lineLength" : 80,
-  "maximumBlankLines" : 1,
-  "prioritizeKeepingFunctionOutputTogether" : false,
-  "respectsExistingLineBreaks" : true,
-  "rules" : {
-    "AllPublicDeclarationsHaveDocumentation" : true,
-    "AlwaysUseLowerCamelCase" : false,
-    "AmbiguousTrailingClosureOverload" : true,
-    "BeginDocumentationCommentWithOneLineSummary" : false,
-    "DoNotUseSemicolons" : true,
-    "DontRepeatTypeInStaticProperties" : false,
-    "FileScopedDeclarationPrivacy" : true,
-    "FullyIndirectEnum" : true,
-    "GroupNumericLiterals" : true,
-    "IdentifiersMustBeASCII" : true,
-    "NeverForceUnwrap" : false,
-    "NeverUseForceTry" : false,
-    "NeverUseImplicitlyUnwrappedOptionals" : false,
-    "NoAccessLevelOnExtensionDeclaration" : false,
-    "NoAssignmentInExpressions" : true,
-    "NoBlockComments" : true,
-    "NoCasesWithOnlyFallthrough" : true,
-    "NoEmptyTrailingClosureParentheses" : true,
-    "NoLabelsInCasePatterns" : false,
-    "NoLeadingUnderscores" : false,
-    "NoParensAroundConditions" : true,
-    "NoVoidReturnOnFunctionSignature" : true,
-    "OneCasePerLine" : true,
-    "OneVariableDeclarationPerLine" : true,
-    "OnlyOneTrailingClosureArgument" : true,
-    "OrderedImports" : false,
-    "ReturnVoidInsteadOfEmptyTuple" : true,
-    "UseEarlyExits" : false,
-    "UseLetInEveryBoundCaseVariable" : false,
-    "UseShorthandTypeNames" : true,
-    "UseSingleLinePropertyGetter" : false,
-    "UseSynthesizedInitializer" : true,
-    "UseTripleSlashForDocumentationComments" : true,
-    "UseWhereClausesInForLoops" : false,
-    "ValidateDocumentationComments" : true
-  },
-  "spacesAroundRangeFormationOperators" : false,
-  "tabWidth" : 4,
-  "version" : 1
+  "indentConditionalCompilationBlocks": false,
+  "indentSwitchCaseLabels": false,
+  "lineBreakAroundMultilineExpressionChainComponents": true,
+  "lineBreakBeforeControlFlowKeywords": true,
+  "lineBreakBeforeEachArgument": true,
+  "lineBreakBeforeEachGenericRequirement": true,
+  "prioritizeKeepingFunctionOutputTogether": false,
+  "respectsExistingLineBreaks": true,
+  "spacesAroundRangeFormationOperators": false,
+  "multiElementCollectionTrailingCommas": true,
+  "rules": {
+    "AllPublicDeclarationsHaveDocumentation": false,
+    "AlwaysUseLiteralForEmptyCollectionInit": true,
+    "AlwaysUseLowerCamelCase": false,
+    "AmbiguousTrailingClosureOverload": true,
+    "BeginDocumentationCommentWithOneLineSummary": true,
+    "DoNotUseSemicolons": true,
+    "DontRepeatTypeInStaticProperties": true,
+    "FileScopedDeclarationPrivacy": true,
+    "FullyIndirectEnum": true,
+    "GroupNumericLiterals": true,
+    "IdentifiersMustBeASCII": true,
+    "NeverForceUnwrap": false,
+    "NeverUseForceTry": true,
+    "NeverUseImplicitlyUnwrappedOptionals": true,
+    "NoAccessLevelOnExtensionDeclaration": true,
+    "NoAssignmentInExpressions": true,
+    "NoBlockComments": true,
+    "NoCasesWithOnlyFallthrough": true,
+    "NoEmptyTrailingClosureParentheses": true,
+    "NoLabelsInCasePatterns": true,
+    "NoLeadingUnderscores": true,
+    "NoParensAroundConditions": true,
+    "NoPlaygroundLiterals": true,
+    "NoVoidReturnOnFunctionSignature": true,
+    "OmitExplicitReturns": true,
+    "OneCasePerLine": true,
+    "OneVariableDeclarationPerLine": true,
+    "OnlyOneTrailingClosureArgument": true,
+    "OrderedImports": true,
+    "ReplaceForEachWithForLoop": true,
+    "ReturnVoidInsteadOfEmptyTuple": true,
+    "TypeNamesShouldBeCapitalized": true,
+    "UseEarlyExits": true,
+    "UseLetInEveryBoundCaseVariable": true,
+    "UseShorthandTypeNames": true,
+    "UseSingleLinePropertyGetter": true,
+    "UseSynthesizedInitializer": true,
+    "UseTripleSlashForDocumentationComments": true,
+    "UseWhereClausesInForLoops": true,
+    "ValidateDocumentationComments": true
+  }
 }
