+++ Package.resolved
@@ -14,17 +14,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/feather-framework/feather-component",
       "state" : {
-        "revision" : "d8373cf9ab5db3e7b21fbcb8e6d1d5718b3bb73b",
-        "version" : "0.4.0"
-      }
-    },
-    {
-      "identity" : "feather-relational-database",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/feather-framework/feather-relational-database",
-      "state" : {
-        "revision" : "fd4fc17854f3a17fe33b1a80c02615bd9da34a7a",
-        "version" : "0.2.0"
+        "revision" : "c209d87352d27afc124422db11c9ca2bb0f7a376",
+        "version" : "0.5.0"
       }
     },
     {
@@ -50,8 +41,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/vapor/sqlite-nio.git",
       "state" : {
-        "revision" : "f3164216b0161f63e609e75f9a28a6f542dd9123",
-        "version" : "1.8.6"
+        "revision" : "2102da7e204fa3b6eeac09a2d01b7aaa1f66c39a",
+        "version" : "1.8.7"
       }
     },
     {
@@ -61,15 +52,6 @@
       "state" : {
         "revision" : "f6919dfc309e7f1b56224378b11e28bab5bccc42",
         "version" : "1.2.0"
-      }
-    },
-    {
-      "identity" : "swift-async-algorithms",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/apple/swift-async-algorithms.git",
-      "state" : {
-        "revision" : "da4e36f86544cdf733a40d59b3a2267e3a7bbf36",
-        "version" : "1.0.0"
       }
     },
     {
@@ -100,12 +82,21 @@
       }
     },
     {
+      "identity" : "swift-nanoid",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/binarybirds/swift-nanoid",
+      "state" : {
+        "revision" : "a7ec5dde163e607ea5fbd473eb5a58908d043eed",
+        "version" : "1.0.0"
+      }
+    },
+    {
       "identity" : "swift-nio",
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/apple/swift-nio.git",
       "state" : {
-        "revision" : "635b2589494c97e48c62514bc8b37ced762e0a62",
-        "version" : "2.63.0"
+        "revision" : "fc63f0cf4e55a4597407a9fc95b16a2bc44b4982",
+        "version" : "2.64.0"
       }
     },
     {
@@ -115,15 +106,6 @@
       "state" : {
         "revision" : "0a5bc04095a675662cf24757cc0640aa2204253b",
         "version" : "1.0.2"
-      }
-    },
-    {
-      "identity" : "swift-service-lifecycle",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/swift-server/swift-service-lifecycle",
-      "state" : {
-        "revision" : "55f45e39dd23c6cad82d4d529e22961cb5e493aa",
-        "version" : "2.4.0"
       }
     },
     {
