+++ README.md
@@ -1,6 +1,6 @@
-# Feather Relational Database Driver SQLite
+# Feather  Database Driver SQLite
 
-SQLite driver for the abstract feather relational-database component for Feather CMS.
+SQLite driver for the abstract feather-database component for Feather CMS.
 
 ## Getting started
 
@@ -13,16 +13,16 @@ Use at your own risk.
 To add a dependency on the package, declare it in your `Package.swift`:
 
 ```swift
-.package(url: "https://github.com/feather-framework/feather-relational-database-driver-sqlite", .upToNextMinor(from: "0.2.0")),
+.package(url: "https://github.com/feather-framework/feather-database-driver-sqlite", .upToNextMinor(from: "0.2.0")),
 ```
 
 and to your application target, add `FeatherSQLDatabase` to your dependencies:
 
 ```swift
-.product(name: "FeatherRelationalDatabaseDriverSQLite", package: "feather-relational-database-driver-sqlite")
+.product(name: "FeatherDatabaseDriverSQLite", package: "feather-database-driver-sqlite")
 ```
 
-Example `Package.swift` file with `FeatherRelationalDatabaseDriverSQLite` as a dependency:
+Example `Package.swift` file with `FeatherDatabaseDriverSQLite` as a dependency:
 
 ```swift
 // swift-tools-version:5.9
@@ -31,11 +31,11 @@ import PackageDescription
 let package = Package(
     name: "my-application",
     dependencies: [
-        .package(url: "https://github.com/feather-framework/feather-relational-database-driver-sqlite", .upToNextMinor(from: "0.2.0")),
+        .package(url: "https://github.com/feather-framework/feather-database-driver-sqlite", .upToNextMinor(from: "0.2.0")),
     ],
     targets: [
         .target(name: "MyApplication", dependencies: [
-            .product(name: "FeatherRelationalDatabaseDriverSQLite", package: "feather-relational-database-driver-sqlite")
+            .product(name: "FeatherDatabaseDriverSQLite", package: "feather-database-driver-sqlite")
         ]),
         .testTarget(name: "MyApplicationTests", dependencies: [
             .target(name: "MyApplication"),
